Extract Windows1250 to its own file
